requirements.txt
tornado
pycrypto
git+https://github.com/mynameisfiber/secret-sharing.git
